+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Non-breaking changes
 
+* Add `BRIDGE_TO_CCIP` action to `BridgeSettlerBase` for Chainlink CCIP cross-chain token transfers
 * Deploy Settler to Tempo mainnet chain
 * Add Thena UniV3 Algebra style fork to Bnb with fork ID 37
 * EkuboV2 VIP actions `EKUBO_VIP` and `METATXN_EKUBO_VIP` are not